fixes after review lesson-2-hometask
