+++ lesson-4/hometask-quiz/src/main/java/io/github/maxixcom/otus/quiz/Main.java
@@ -1,5 +1,6 @@
 package io.github.maxixcom.otus.quiz;
 
+import io.github.maxixcom.otus.quiz.config.AppConfig;
 import io.github.maxixcom.otus.quiz.service.ApplicationRunner;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
