+++ lesson-4/hometask-quiz/src/main/java/io/github/maxixcom/otus/quiz/domain/Question.java
@@ -2,18 +2,14 @@ package io.github.maxixcom.otus.quiz.domain;
 
 public abstract class Question {
     private final String content;
-    private final Answer correctAnswer;
 
-    protected Question(String content, Answer correctAnswer) {
+    protected Question(String content) {
         this.content = content;
-        this.correctAnswer = correctAnswer;
     }
 
     public String getContent() {
         return this.content;
     }
 
-    public Answer getCorrectAnswer() {
-        return correctAnswer;
-    }
+    abstract public Answer getCorrectAnswer();
 }
