+++ lesson-4/hometask-quiz/src/main/java/io/github/maxixcom/otus/quiz/domain/QuestionChoice.java
@@ -4,13 +4,19 @@ import java.util.List;
 
 public class QuestionChoice extends Question {
     private final List<Answer> options;
+    private final Answer correctAnswer;
 
     public QuestionChoice(String content, Answer correctAnswer, List<Answer> options) {
-        super(content, correctAnswer);
+        super(content);
+        this.correctAnswer = correctAnswer;
         this.options = options;
     }
 
     public List<Answer> getOptions() {
         return options;
+    }
+
+    public Answer getCorrectAnswer() {
+        return correctAnswer;
     }
 }
