+++ lesson-4/hometask-quiz/src/main/java/io/github/maxixcom/otus/quiz/domain/QuestionGeneral.java
@@ -1,7 +1,14 @@
 package io.github.maxixcom.otus.quiz.domain;
 
 public class QuestionGeneral extends Question {
+    private final Answer correctAnswer;
+
     public QuestionGeneral(String content, Answer correctAnswer) {
-        super(content, correctAnswer);
+        super(content);
+        this.correctAnswer = correctAnswer;
+    }
+
+    public Answer getCorrectAnswer() {
+        return correctAnswer;
     }
 }
